feat(condo): DOMA-8507 Try to fix "module not found" error
