+++ packages/clients/address-service-client/index.js
@@ -1,20 +1,20 @@
 const get = require('lodash/get')
 
-const { AddressServiceClient: MockedClient } = require('@open-condo/clients/address-service-client/__mocks__/AddressServiceClient')
 const { AddressServiceClient } = require('@open-condo/clients/address-service-client/AddressServiceClient')
+const { MockedAddressServiceClient } = require('@open-condo/clients/address-service-client/MockedAddressServiceClient')
 const conf = require('@open-condo/config')
 
 let instance
 
 /**
- * @returns {AddressServiceClient}
+ * @returns {AddressServiceClient|MockedAddressServiceClient}
  */
 function createInstance () {
     const addressServiceUrl = get(conf, 'ADDRESS_SERVICE_DOMAIN')
 
     if (!instance) {
         if (get(conf, 'ADDRESS_SERVICE_CLIENT_MODE') === 'fake') {
-            instance = new MockedClient(addressServiceUrl)
+            instance = new MockedAddressServiceClient(addressServiceUrl)
         } else {
             instance = new AddressServiceClient(addressServiceUrl)
         }
